Finished
I had a previous project done but when making the code for misssion demolition something happened to the project that made it so the scripts were showing up as not exisiting. i tried making new scripts but it didn't solve the issue. I had to start a new one and just rushed through it